MDL-36030 grade: Fix display totals excluding hidden items for students.
In the original issue MDL-21218 the handling of course / category / manual items wasn't updated to handle excluding
hidden items. So this displays incorrectly to the student when gradebook course settings are set the course to show the
total excluding hidden grades.

This is fixed by allowing the course / category / manual item to be process normally.

Noting: Manual grade items with calculations include hidden items grades.